app/requirements.txt
streamlit==1.51.0
tensorflow==2.20.0
Pillow==12.0.0
numpy==2.3.5